refactor: errorHandler extracted to own service file
