+++ src/web/src/services/commandApi.ts
@@ -1,4 +1,5 @@
 import axios from "axios";
+import { apiErrorHandler } from "./errorHandler";
 
 export const commandApi = {
   getCommand: async (leafUrl: string): Promise<any> => {
@@ -104,26 +105,5 @@ export const commandApi = {
     } catch (err: any) {
       apiErrorHandler.handleApiError(err, "Failed to create subresource");
     }
-  },
-} as const;
-
-export const apiErrorHandler = {
-  getErrorMessage: (err: any): string => {
-    if (err.response?.data?.message) {
-      const data = err.response.data;
-      const details = data.details ? `: ${JSON.stringify(data.details)}` : "";
-      return `ResponseError: ${data.message}${details}`;
-    }
-    return "An unexpected error occurred";
-  },
-
-  isHttpError: (err: any, statusCode: number): boolean => {
-    return err.response?.status === statusCode;
-  },
-
-  handleApiError: (err: any, context: string = ""): never => {
-    console.error(context, err);
-    const message = apiErrorHandler.getErrorMessage(err);
-    throw new Error(message);
   },
 } as const;
