+++ src/web/src/services/index.ts
@@ -1,4 +1,5 @@
 export { workspaceApi, type Workspace, type CreateWorkspaceData, type ClientConfig } from "./workspaceApi";
 export { specsApi, specsHelper, type Plane, type Resource } from "./specsApi";
-export { commandApi, apiErrorHandler } from "./commandApi";
+export { commandApi } from "./commandApi";
+export { apiErrorHandler } from "./errorHandler";
 export { cliApi } from "./cliApi";
